+++ cluster/local/integration_tests_compute.sh
@@ -462,47 +462,12 @@ EOF
   echo "${INSTALL_RESOURCE_YAML}" | "${KUBECTL}" apply -f -
 
   echo_step "waiting for nic CR to be ready & synced"
+  sleep 10s
   kubectl wait --for=condition=ready nics/example --timeout 120s
   kubectl wait --for=condition=synced nics/example --timeout 120s
 
   echo_step "get nic CR"
   kubectl get nics
-
-  echo_step "update nic CR"
-  INSTALL_RESOURCE_YAML="$(
-    cat <<EOF
-apiVersion: compute.ionoscloud.crossplane.io/v1alpha1
-kind: Nic
-metadata:
-  name: example
-spec:
-  forProvider:
-    name: exampleNic
-    dhcp: true
-    ipsConfigs:
-      ipsBlockConfigs:
-        - ipBlockIdRef:
-            name: example
-    firewallActive: true
-    datacenterConfig:
-      datacenterIdRef:
-        name: example
-    serverConfig:
-      serverIdRef:
-        name: example
-    lanConfig:
-      lanIdRef:
-        name: example
-  providerConfigRef:
-    name: example
-EOF
-  )"
-
-  echo "${INSTALL_RESOURCE_YAML}" | "${KUBECTL}" apply -f -
-
-  echo_step "waiting for nic CR to be ready & synced"
-  kubectl wait --for=condition=ready nics/example --timeout 120s
-  kubectl wait --for=condition=synced nics/example --timeout 120s
 }
 
 function nic_tests_cleanup() {