+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [0.1.0-alpha.3] (upcoming release)
 
 - **Breaking Changes**:
-    - updated `spec.forProvider.ips` field from **Nic Managed Resource** to `spec.forProvider.ipsConfig` being able to
+    - updated `spec.forProvider.ips` field from **Nic Managed Resource** to `spec.forProvider.ipsConfigs` being able to
       set IPs directly or via references and indexes of the IPBlocks
     - updated `spec.forProvider.ip` field from **IPFailover Managed Resource** to `spec.forProvider.ipConfig` being able
       to set the required IP directly or via reference and index to an IPBlock